requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
discord.py==2.6.4
dotenv==0.9.9
exceptiongroup==1.3.1
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
playwright==1.58.0
propcache==0.4.1
pyee==13.0.0
python-dotenv==1.2.1
typing_extensions==4.15.0
yarl==1.22.0
